requirements.txt
Django==5.2.3
djangorestframework==3.15.2
python-dotenv==1.0.1
